+++ src/functionalTest/java/uk/gov/hmcts/reform/wataskconfigurationapi/PostTaskConfigurationTest.java
@@ -1,29 +1,14 @@
 package uk.gov.hmcts.reform.wataskconfigurationapi;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import io.restassured.response.Response;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.After;
 import org.junit.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
-import uk.gov.hmcts.reform.authorisation.generators.AuthTokenGenerator;
-import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.ccd.client.model.Event;
-import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.idam.IdamSystemTokenGenerator;
-import uk.gov.hmcts.reform.wataskconfigurationapi.auth.idam.entities.UserInfo;
 import uk.gov.hmcts.reform.wataskconfigurationapi.controllers.request.ConfigureTaskRequest;
-import uk.gov.hmcts.reform.wataskconfigurationapi.utils.CreateTaskMessage;
-import uk.gov.hmcts.reform.wataskconfigurationapi.utils.RoleAssignmentHelper;
+import uk.gov.hmcts.reform.wataskconfigurationapi.services.CreateTaskMessage;
 
-import java.io.IOException;
 import java.util.Map;
-import java.util.Objects;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicReference;
 
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.CoreMatchers.nullValue;
@@ -38,21 +23,9 @@ public class PostTaskConfigurationTest extends SpringBootFunctionalBaseTest {
 
     private static final String ENDPOINT_BEING_TESTED = "task/{task-id}/configuration";
 
-    @Autowired
-    private AuthTokenGenerator serviceAuthTokenGenerator;
-
-    @Autowired
-    private IdamSystemTokenGenerator systemTokenGenerator;
-
-    @Autowired
-    private CoreCaseDataApi coreCaseDataApi;
-
-    @Autowired
-    private RoleAssignmentHelper roleAssignmentHelper;
-
     private String taskId;
-    private CreateTaskMessage createTaskMessage;
     private String caseId;
+    private CreateTaskMessage createTaskMessage;
 
     @After
     public void cleanUp() {
@@ -138,106 +111,4 @@ public class PostTaskConfigurationTest extends SpringBootFunctionalBaseTest {
             .body("configuration_variables.taskSystem", equalTo("SELF"));
     }
 
-    private String createTask(CreateTaskMessage createTaskMessage) {
-
-        Response camundaResult = camundaApiActions.post(
-            "/message",
-            createTaskMessage,
-            authorizationHeadersProvider.getServiceAuthorizationHeader()
-        );
-
-        camundaResult.then().assertThat()
-            .statusCode(HttpStatus.NO_CONTENT.value());
-
-        Object taskName = createTaskMessage.getProcessVariables().get("name").getValue();
-
-        String filter = "?processVariables=" + "caseId_eq_" + createTaskMessage.getCaseId();
-
-        AtomicReference<String> response = getTaskId(taskName, filter);
-        
-        return response.get();
-        
-    }
-
-    private String createCcdCase() throws IOException {
-        String userToken = systemTokenGenerator.generate();
-        UserInfo userInfo = systemTokenGenerator.getUserInfo(userToken);
-        String serviceToken = serviceAuthTokenGenerator.generate();
-        StartEventResponse startCase = coreCaseDataApi.startForCaseworker(
-            userToken,
-            serviceToken,
-            userInfo.getUid(),
-            "IA",
-            "Asylum",
-            "startAppeal"
-        );
-        String caseData = new String(
-            (Objects.requireNonNull(Thread.currentThread().getContextClassLoader()
-                .getResourceAsStream("case_data.json"))).readAllBytes()
-        );
-        Map data = new ObjectMapper().readValue(caseData, Map.class);
-        CaseDataContent caseDataContent = CaseDataContent.builder()
-            .eventToken(startCase.getToken())
-            .event(Event.builder()
-                .id(startCase.getEventId())
-                .summary("summary")
-                .description("description")
-                .build())
-            .data(data)
-            .build();
-
-        CaseDetails caseDetails = coreCaseDataApi.submitForCaseworker(
-            userToken,
-            serviceToken,
-            userInfo.getUid(),
-            "IA",
-            "Asylum",
-            true,
-            caseDataContent
-        );
-
-        log.info("Created case [" + caseDetails.getId() + "]");
-
-        StartEventResponse submitCase = coreCaseDataApi.startEventForCaseWorker(
-            userToken,
-            serviceToken,
-            userInfo.getUid(),
-            "IA",
-            "Asylum",
-            caseDetails.getId().toString(),
-            "submitAppeal"
-        );
-
-        CaseDataContent submitCaseDataContent = CaseDataContent.builder()
-            .eventToken(submitCase.getToken())
-            .event(Event.builder()
-                .id(submitCase.getEventId())
-                .summary("summary")
-                .description("description")
-                .build())
-            .data(data)
-            .build();
-        coreCaseDataApi.submitEventForCaseWorker(
-            userToken,
-            serviceToken,
-            userInfo.getUid(),
-            "IA",
-            "Asylum",
-            caseDetails.getId().toString(),
-            true,
-            submitCaseDataContent
-        );
-        log.info("Submitted case [" + caseDetails.getId() + "]");
-        //Added wait as there seems to be a delay while retrieving the case.
-        waitSeconds(2);
-        return caseDetails.getId().toString();
-    }
-
-    private void waitSeconds(int seconds) {
-        try {
-            TimeUnit.SECONDS.sleep(seconds);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-    }
 }
