+++ src/testUtils/java/uk/gov/hmcts/reform/wataskconfigurationapi/services/RoleAssignmentHelper.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.wataskconfigurationapi.utils;
+package uk.gov.hmcts.reform.wataskconfigurationapi.services;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
